Remove unused tracker detector instance
